requirements.txt
3to2==1.1.1
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
applib==1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
Babel==2.9.1
beautifulsoup4==4.12.2
blinker==1.7.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
cmdln==2.0.0
CodeIntel==2.0.0
cryptography==3.4.8
Cython==3.0.6
dataclasses-json==0.6.2
decorator==4.4.2
deepface==0.0.79
defusedxml==0.7.1
Django==4.2.5
djangorestframework==3.14.0
dlib==19.24.2
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.5
environs==9.5.0
esprima==4.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
freezegun==0.3.15
frozenlist==1.4.0
gast==0.5.4
gdown==4.7.1
gevent==22.10.2
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
idna==2.10
Inflector==3.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.15.0
langchain==0.0.340
langsmith==0.0.66
libclang==16.0.6
libclang-py3==3.9.0
libsass==0.22.0
lxml==4.9.2
Markdown==3.5.1
MarkupSafe==2.1.2
marshmallow==3.20.1
minio==7.2.0
ml-dtypes==0.2.0
mtcnn==0.1.1
multidict==6.0.4
mypy-extensions==1.0.0
num2words==0.5.9
numpy==1.26.2
oauthlib==3.2.2
ofxparse==0.21
openai==0.28.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging @ file:///private/tmp/python-packaging-20231001-5984-iihu8s/packaging-23.2
pandas==2.1.3
passlib==1.7.4
Pillow==9.4.0
podman-compose==1.0.6
polib==1.1.0
protobuf==4.23.4
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///private/tmp/pycparser-20230831-6050-10bkury/pycparser-2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
pymilvus==2.3.3
pyOpenSSL==20.0.1
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==2.12.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ldap==3.4.0
python-multipart==0.0.6
python-stdnum==1.16
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
qrcode==6.1
regex==2023.10.3
reportlab==3.6.12
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retina-face==0.0.13
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.5
uvicorn==0.24.0.post1
validators==0.22.0
vobject==0.9.6.1
weaviate-client==3.25.3
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarl==1.9.2
zeep==4.0.0
zope.cachedescriptors==5.0
zope.event==5.0
zope.interface==6.1
